AlterColumnOperation => Fix the "set default" statement. Apply to all types except blob (either text & binary)
RenameIndexOperation => "rename index" do not exist, must drop it, and regenerate the ddl to recreate it. Not completely tested.

RenameTableOperation => Commented version of the solution, not yet usable.

FbDropDatabaseOperation => Implemented, won't work if not sysdba maybe a test to be added.

DropIndexOperation, DropForeignKeyOperation => Fixed syntax

RenameColumnOperation => Mysql version replaced with Fb